Transfer services and interfaces to their proper directories for more organized code structure
