+++ TypeScript/app/models/gilded-rose.ts
@@ -1,0 +1,6 @@
+import { IItem } from '@/models/item'
+
+export interface IGildedRose {
+  items: Array<IItem>
+  updateQuality(): void
+}
