+++ TypeScript/app/models/index.ts
@@ -1,0 +1,2 @@
+export * from './item'
+export * from './gilded-rose'
