+++ TypeScript/app/services/index.ts
@@ -1,0 +1,2 @@
+export * from './gilded-rose'
+export * from './item'
