+++ TypeScript/app/services/item.ts
@@ -1,0 +1,13 @@
+import { IItem } from '@/models'
+
+export class Item implements IItem {
+  name: string
+  sellIn: number
+  quality: number
+
+  constructor(name: string, sellIn: number, quality: number) {
+    this.name = name
+    this.sellIn = sellIn
+    this.quality = quality
+  }
+}
